Refactor extension progress display and remove unused progress pill component.
