--- app/components/course-page/configure-extensions-modal/extension-progress-pill.hbs
@@ -1,3 +1,0 @@
-{{#if this.progressPillColor}}
-  <Pill @color={{this.progressPillColor}} data-test-extension-progress-pill>{{this.progressPillText}}</Pill>
-{{/if}}
